Socket Bug Fixes
The protocol for the handshake between client and server required a passing of a websocket. A Proxy Pass on the Apache2 server with Rewrite Engine to handle the HTTP Headers which were throwing a bad response header 400 Error.
On the client side, I had to set upgrades to false:  Initially, Socket.IO.client will send an XHR and JSONP request as per its protocol, only updating to webSocket if its available. By disabling this, the initial connect speed is increased and an error 400 is no longer occuring. Client Connect/Disconnect is being logged on server side with this approach.